chore: rename LLM provider select helpers
