+++ platform/frontend/src/app/llm/(costs)/optimization-rules/_parts/rule.tsx
@@ -12,7 +12,7 @@ import {
   LlmModelPicker,
   type ModelPricing,
 } from "@/components/llm-model-picker";
-import { LlmProviderSelectItems } from "@/components/llm-provider-options";
+import { LlmProviderSelectItems } from "@/components/llm-provider-select-items";
 import { WithPermissions } from "@/components/roles/with-permissions";
 import { Badge } from "@/components/ui/badge";
 import { Button } from "@/components/ui/button";
